+++ src/Services/index.js
@@ -2,7 +2,6 @@ export * as authService from "./authService";
 export * as locationService from "./locationService";
 export * as feedbackService from "./feedbackService";
 export * as heartRateService from "./heartRateService";
-export * as localProgramImportService from "./localProgramImportService";
 export * as notificationService from "./notificationService";
 export * as ownWorkoutPostService from "./ownWorkoutPostService";
 export * as programService from "./programService";